# apps/backend/api/swag-docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_SURF-Innovatie_MORIS_internal_api_userdto.Request:
    properties:
      password:
        type: string
      person_id:
        type: string
    type: object
  github_com_SURF-Innovatie_MORIS_internal_api_userdto.Response:
    properties:
      email:
        type: string
      familyName:
        type: string
      givenName:
        type: string
      id:
        type: string
      name:
        type: string
      orcid:
        type: string
      person_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: MORIS
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MORIS
  version: "1.0"
paths:
...
  /users:
    delete:
      consumes:
      - application/json
      description: Deletes a user by its ID, provided as the `id` query parameter.
      parameters:
      - description: User ID (UUID)
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: no content
          schema:
            type: string
        "400":
          description: invalid id
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Delete a user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Retrieves a single user by its ID, provided as the `id` query parameter.
      parameters:
      - description: User ID (UUID)
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_SURF-Innovatie_MORIS_internal_api_userdto.Response'
        "400":
          description: invalid id
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Get a user by ID
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Creates a new user account for an existing person using the provided
        person ID and password.
      parameters:
      - description: User creation payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_SURF-Innovatie_MORIS_internal_api_userdto.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_SURF-Innovatie_MORIS_internal_api_userdto.Response'
        "400":
          description: invalid request body or missing person ID
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Create a new user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Updates an existing user's person reference and/or password based
        on the given ID and request body.
      parameters:
      - description: User ID (UUID)
        in: query
        name: id
        required: true
        type: string
      - description: User update payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_SURF-Innovatie_MORIS_internal_api_userdto.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_SURF-Innovatie_MORIS_internal_api_userdto.Response'
        "400":
          description: invalid id or request body
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Update a user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" then a space and your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"